vos: Remove client connection rx_ServiceOf() calls
Since the original IBM import, the 'vos dump' command has logic to
adjust the rx dead time for the rx_service of our client conns.  But
client conns by definition do not have an rx_service associated with
them, only a service id.  Therefore this code has "always" been
harmlessly superfluous (at least, "it was like that when we got it").

Later, with commit 1f3990e5eafb vos-getsize-20030619, this logic was
cargo-culted into the then-new 'vos size' command as well.

Remove both instances of trying to update a non-existent service.

No functional change should be incurred by this commit.

(This was discovered during an audit of rx_service timeouts and
rx_ServiceOf usage.)

[mmeffie: update commit message]

Change-Id: Icffd5860be3c38c78cc16d7606d760364c5fa3b4
Reviewed-on: https://gerrit.openafs.org/15726
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>