afs: Refactor afs_Analyze to use common exit label
The function afs_Analyze() currently has a common exit label `out` that
is not used consistently.

Refactor afs_Analyze() to branch to `out` instead of returning.  Update
the code at `out` to only call afs_PutConn() if aconn is not NULL.

Move the call to afs_FinalizeReq() so it's earlier in the code path to
simplify early returns.

Replace a cast of 0 as a pointer with NULL.

Minor white space cleanup.

Change-Id: I552d84dd653a483774551dad1c05d7a692ac9d14
Reviewed-on: https://gerrit.openafs.org/15745
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>