doc: Split CODING into separate documents
Split the CODING file into more management sized documents, and move
them to a new home under doc/process, which is the new location
for process oriented documention.

This was a mechanical split so the file contents are unchanged. sed was
used to split the CODING file by line numbers.

    mkdir -p doc/process
    sed -n '1,198p'   CODING >doc/process/code-style-guide.md
    sed -n '200,295p' CODING >doc/process/howto-build-openafs.md
    sed -n '297,319p' CODING >doc/process/backporting-guide.md
    sed -n '321,509p' CODING >doc/process/code-review-guide.md
    sed -n '511,$p'   CODING >doc/process/compiler-warnings.md

Change-Id: I44bc69c23f15cc6319046ea7d5908423db80879c
Reviewed-on: https://gerrit.openafs.org/15861
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: BuildBot <buildbot@rampaginggeek.com>