doc: Add OpenAFS Contributor Guide
Add guidelines for OpenAFS contributors.

The majority of the submitting-changes and commit-messages guides were
written by Andrew Deason, as well as text for the other guides.

Corrections to the guides and formatting fixes were provided by Sahil
Siddiq and Cheyenne Wills.

The text from the CODING file was moved to this series of documents.
Authors and contributors of the CODING file (and the original
README.devel file) were:

* Anders Kaseorg
* Andrew Deason
* Benjamin Kaduk
* Cheyenne Wills
* Daria Phoebe Brashear
* Garrett Wollman
* Jim Rees
* Marcio Barbosa
* Michael Meffie
* Nathan Neulinger
* Russ Allbery
* Simon Wilkinson

Change-Id: I8827ca020a7cb50304d7c4ece3ac0572f41f6a6a
Reviewed-on: https://gerrit.openafs.org/15864
Reviewed-by: Gaurav Saxena <gsaxena@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>