stds.h: Fix comment/indentation for static_inline
The comment proceeding the block of defines for the static_inline
is incorrect.  In addition the indentation is missing for the
preprocessor statements within this block.

The commits:
    "printf-sanity-20090317" (3553442536)
    "printf-sanity-20090318" (be4d4076f2)
introduced the functions afs_cast_int32() and afs_cast_uint32() and the
associated comment for these functions.

Later the commits:
    "static-inline-macros-20090319" (101399171b)
    "Rename printf cast helpers and clean up format string warnings"
    (ac3e0ed031)
moved afs_cast_int32() and afs_cast_uint32() and later renamed them, but
failed to relocate the associated comment.

Relocate the existing comment for the casting functions and add
an appropriate comment for the "static_inline" macro.

Fix the preprocessor indentation for this block.

There are no functional changes with this commit

Change-Id: Ie49f2dd4a8f18c7855e400f32aa39f49e1a15fb1
Reviewed-on: https://gerrit.openafs.org/15767
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>