xdr: Set 0 _len with NULL _val
When passing a struct to an RPC with an OUT opaque or array arg, a few
callers give a NULL _val (to indicate that xdr should allocate the
memory for us), but leave _len uninitialized. This is okay, because xdr
ignores the _len when _val is NULL, but set _len to 0 in these cases
just in case, and to be more consistent with other callers.

Reviewed-on: https://gerrit.openafs.org/16240
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry picked from commit 22f87bcb04d9127f8c8c11ebb3b8cd96028c3d4a)

Change-Id: I45820542940ce66e7b392f060cd4b867c8f2ee2a
Reviewed-on: https://gerrit.openafs.org/16267
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>