xdr: Set 0 _len with NULL _val
When passing a struct to an RPC with an OUT opaque or array arg, a few
callers give a NULL _val (to indicate that xdr should allocate the
memory for us), but leave _len uninitialized. This is okay, because xdr
ignores the _len when _val is NULL, but set _len to 0 in these cases
just in case, and to be more consistent with other callers.

Change-Id: I33d541fbdaaf7f8c6cd724ddd7140c0ebd1fa1ae
Reviewed-on: https://gerrit.openafs.org/16240
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>