doc: Use ln force option to install man pages
The makefile target to install man pages tests for the existence of the
man page symlinks being installed and skips creating the symlinks that
already exist.

Instead, use the `ln` -f (force) option to recreate the symlinks, even
if they already exist. This ensures the links are correct and slightly
simplifies the man-page install recipe in the Makefile.

The `ln` -f (force) option is required by POSIX so should be available
on all the platforms we build on.

Change-Id: I79ebbcfb5d927232498e26607504eefe84164596
Reviewed-on: https://gerrit.openafs.org/16319
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>