LINUX: Add newline to end of osi_misc.c
Since commit 840cf3b62891 (afs: Introduce afs_kill_pending()),
osi_misc.c has lacked a newline at the end of the file. This triggers a
warning on some old compilers, such as gcc 4.1 on RHEL5, which breaks
the build when using --enable-checking:

    [...]/src/libafs/MODLOAD-2.6.18-419.el5-SP/osi_misc.c:205:2: error: no newline at end of file

Add a newline to get rid of the warning/error.

Change-Id: Id5a3a3f36c4f51b36426b2fe78ac5afb1f4fb6b3
Reviewed-on: https://gerrit.openafs.org/15775
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>