butm: Remove duplicate library arguments
The test_ftm and butm_test programs link against the libbutm.a library.
Currently those targets depend on libbutm.a but the link command also
includes the ${TOP_LIBDIR}/libbutm.a which is made by the install
target.

This can lead to errors during a parallel build:

    gcc ... -o test_ftm test_ftm.o libbutm.a ... <top_dir>/lib/libbutm.a ...
    gcc: error: <top_dir>/lib/libbutm.a: No such file or directory

Remove the ${TOP_LIBDIR}/libbutm.a from the LIBS macro, so the test_ftm
and butm_test programs are built from the libbutm.a in the local
directory, which those programs depend on.

Change-Id: I695b0bbff4053804c514b361dca4da49e623a2a3
Reviewed-on: https://gerrit.openafs.org/15781
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: Michael Meffie <mmeffie@sinenomine.net>