doc: Sort man page entries in AdminRef
The AdminRef is a collection of OpenAFS man pages in a single document.
Unfortunately, the man pages are not listed in any particular order, but
rather just the order found by reading the unsorted directory entries.

Change the generate-xml script so the man pages are in sorted order in
the generated AdminRef document.  Instead of writing the man page
references after processing each page, save each entry in a list, and
then sort the list after all the pages have been processed.

Also, check the exit code of the pod2refentry script so errors are not
ignored while generating the AdminRef document.

Change-Id: I624485e4efd1fb0c08642b379a12c946f5793336
Reviewed-on: https://gerrit.openafs.org/15827
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>