bosserver: use normal asprintf idiom
We generally tend to explicitly NULL out the buffer argument if
asprintf() returns failure, since it is not guaranteed to have done
so itself.  It is even less clear whether setting the buffer argument
to NULL ourselves before calling asprintf() will be untouched in all
possible failures, so re-zero it after failed calls just to be sure
we don't try to free garbage later.

Change-Id: Ie8995472aeadb8043f5e43f4e6d197a7be1363e0
Reviewed-on: https://gerrit.openafs.org/15832
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>