Remove MIN and MAX macro definitions
A prior commit:
    "opr: replace MIN/MAX macros with opr_min/opr_max"
       (Change I2d7b54193ec91f7ead9c5c5f714d9a8bc7533bf7)
replaced all uses of the MIN and MAX macros with opr_min and opr_max.

As a cleanup and to resolve a failure when building the Linux kernel
module with Linux 6.11, remove all the defines for MIN and MAX.

The Linux 6.11 commit:
    'minmax: make generic MIN() and MAX() macros available everywhere'
    (1a251f52cf)
standardized and consolidated the definitions of the MIN and MAX macros
within the Linux kernel by defining them in an include file that is
widely used already (linux/minmax.h).

With the above Linux commit, the kernel module fails with a redefined
error from the compiler:
    "./include/linux/minmax.h:329: error: "MIN" redefined [-Werror]"

Reviewed-on: https://gerrit.openafs.org/15814
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry picked from commit 8e8ee623d1a21dc8476cde4b6b9d9e292b36691d)

[mmeffie: The MIN/MAX macros are still present in the deprecated afsweb
          component.]

Change-Id: I096c386682afbf7de07f7bb882ab9442cbdeb13a
Reviewed-on: https://gerrit.openafs.org/15854
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>